Add files via upload
This is an AI desktop assistant which takes commands as speech and then does the respective task.
This Python-based desktop assistant serves as a voice-activated virtual companion, capable of performing various tasks and responding to user commands. The assistant utilises speech recognition, text-to-speech, and automation libraries to interpret user input and execute commands. It includes functionalities such as greeting the user, providing information about the time, opening websites (Google and YouTube), playing songs  and taking screenshots. The user interacts with the assistant by issuing voice commands, and the assistant responds with spoken feedback or performs the requested tasks. The project showcases the integration of different Python libraries to create a simple yet versatile desktop assistant.